fix: replace deprecated ui option with ui_config
BREAKING CHANGE: This cannot be updated in a non-breaking fashion, but
the least disruptive route was chosen.

BREAKING CHANGE: If both `ui` and `ui_config.enabled` are set,
`ui_config` takes precedence.  Hence, the change may enable the UI on
machines which had previously set `ui` to false. This is arguably better
than defaulting to `false`, which would disable the UI where it is
supposed to be enabled.

BREAKING CHANGE: Removing the option entirely breaks similarly if users
rely on the formula defaults, since Consul's default is `false` and the
formula's default used to be `true`.

BREAKING CHANGE: The only other way to break less would be to set both
options, but then users would also have override both (which is not
obvious and very annoying) and there would still be no way forward to
when Consul actually removes the deprecated option.